+++ llvm/include/llvm/IR/InstrTypes.h
@@ -1262,6 +1262,9 @@ public:
     return nullptr;
   }
 
+  /// Get memory effects specific to floating-point operations.
+  LLVM_ABI std::optional<MemoryEffects> getFloatingPointMemoryEffects() const;
+
   static bool classof(const Instruction *I) {
     return I->getOpcode() == Instruction::Call ||
            I->getOpcode() == Instruction::Invoke ||